restructure rpcs paths, fix references
