+++ src/core/atomicdex/api/mm2/mm2.client.hpp
@@ -10,8 +10,8 @@
 #include "atomicdex/utilities/cpprestsdk.utilities.hpp"
 #include "rpc.disable.hpp"
 #include "rpc.recover.funds.hpp"
-#include "rpc.hpp"
-#include "rpc2.task.enable_z_coin.cancel.hpp"
+#include "atomicdex/api/mm2/rpc.hpp"
+#include "atomicdex/api/mm2/rpc_v2/rpc2.task.enable_z_coin.cancel.hpp"
 
 namespace atomic_dex::mm2
 {
