+++ src/core/atomicdex/api/mm2/rpc.cpp
@@ -1,6 +1,6 @@
 #include <nlohmann/json.hpp>
 
-#include "rpc.hpp"
+#include "atomicdex/api/mm2/rpc.hpp"
 
 namespace atomic_dex::mm2
 {
