+++ src/core/atomicdex/api/mm2/rpc_v2/rpc2.enable_slp_rpc.cpp
@@ -1,6 +1,6 @@
 #include <nlohmann/json.hpp>
 
-#include "enable_slp_rpc.hpp"
+#include "rpc2.enable_slp_rpc.hpp"
 
 namespace atomic_dex::mm2
 {
