+++ src/core/atomicdex/api/mm2/rpc_v2/rpc2.enable_tendermint_token.hpp
@@ -21,8 +21,8 @@
 
 #include <nlohmann/json_fwd.hpp> //> nlohmann::json
 
-#include "rpc.hpp"
-#include "balance_info.hpp"
+#include "atomicdex/api/mm2/rpc.hpp"
+#include "atomicdex/api/mm2/balance_info.hpp"
 
 namespace atomic_dex::mm2
 {
