+++ src/core/atomicdex/api/mm2/rpc_v2/rpc2.enable_tendermint_with_assets.cpp
