+++ src/core/atomicdex/api/mm2/rpc_v2/rpc2.z_coin_tx_history.hpp
@@ -24,7 +24,7 @@
 #include <nlohmann/json_fwd.hpp>
 
 //! Our Headers
-#include "transaction.data.hpp"
+#include "atomicdex/api/mm2/transaction.data.hpp"
 
 namespace atomic_dex::mm2
 {
