+++ src/core/atomicdex/services/mm2/mm2.service.cpp
@@ -31,9 +31,9 @@
 #include "atomicdex/api/mm2/rpc.enable.hpp"
 #include "atomicdex/api/mm2/rpc.min.volume.hpp"
 #include "atomicdex/api/mm2/rpc.tx.history.hpp"
-#include "atomicdex/api/mm2/rpc2.z_coin_tx_history.hpp"
-#include "atomicdex/api/mm2/rpc2.task.enable_z_coin.init.hpp"
-#include "atomicdex/api/mm2/rpc2.task.enable_z_coin.status.hpp"
+#include "atomicdex/api/mm2/rpc_v2/rpc2.z_coin_tx_history.hpp"
+#include "atomicdex/api/mm2/rpc_v2/rpc2.task.enable_z_coin.init.hpp"
+#include "atomicdex/api/mm2/rpc_v2/rpc2.task.enable_z_coin.status.hpp"
 #include "atomicdex/config/mm2.cfg.hpp"
 #include "atomicdex/config/coins.cfg.hpp"
 #include "atomicdex/constants/dex.constants.hpp"
