+++ src/core/atomicdex/services/mm2/mm2.service.hpp
@@ -35,10 +35,10 @@
 #include "atomicdex/api/mm2/rpc.max.taker.vol.hpp"
 #include "atomicdex/api/mm2/rpc.min.volume.hpp"
 #include "atomicdex/api/mm2/rpc.orderbook.hpp"
-#include "atomicdex/api/mm2/enable_bch_with_tokens_rpc.hpp"
-#include "atomicdex/api/mm2/enable_slp_rpc.hpp"
-#include "atomicdex/api/mm2/rpc2.enable_tendermint_with_assets.hpp"
-#include "atomicdex/api/mm2/rpc2.enable_tendermint_token.hpp"
+#include "atomicdex/api/mm2/rpc_v2/rpc2.enable_bch_with_tokens_rpc.hpp"
+#include "atomicdex/api/mm2/rpc_v2/rpc2.enable_slp_rpc.hpp"
+#include "atomicdex/api/mm2/rpc_v2/rpc2.enable_tendermint_with_assets.hpp"
+#include "atomicdex/api/mm2/rpc_v2/rpc2.enable_tendermint_token.hpp"
 #include "atomicdex/config/raw.mm2.coins.cfg.hpp"
 #include "atomicdex/constants/dex.constants.hpp"
 #include "atomicdex/data/dex/orders.and.swaps.data.hpp"
