+++ src/core/atomicdex/services/price/orderbook.scanner.service.cpp
@@ -15,7 +15,7 @@
  ******************************************************************************/
 
 //! Project Headers
-#include "atomicdex/api/mm2/rpc2.best.orders.hpp"
+#include "atomicdex/api/mm2/rpc_v2/rpc2.best.orders.hpp"
 #include "atomicdex/pages/qt.trading.page.hpp"
 #include "atomicdex/services/mm2/mm2.service.hpp"
 #include "atomicdex/services/price/orderbook.scanner.service.hpp"
